requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pillow>=10.0.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
onnxruntime>=1.15.0
onnx>=1.14.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
pandas>=2.0.0
pytest>=7.4.0
httpx>=0.24.0
python-dotenv>=1.0.0 